+++ src/components/Layout/SplitPaneLayout.tsx
@@ -4,15 +4,15 @@ import { HeadBodyLayout } from "./HeadBodyLayout";
 
 interface ISplitPaneLayoutProps {
   header?: React.ReactElement;
-  leftPane: React.ReactElement;
+  leftPane?: React.ReactElement;
   rightPane?: React.ReactElement;
 }
 
 export const SplitPaneLayout = (props: ISplitPaneLayoutProps) => {
   const { header, rightPane, leftPane } = props;
 
-  if (!rightPane) {
-    return <HeadBodyLayout header={header} body={leftPane} />;
+  if (!leftPane || !rightPane) {
+    return <HeadBodyLayout header={header} body={leftPane || rightPane} />;
   }
 
   return (
