+++ src/containers/NetworkPanel/NetworkDetails/PanelSection.tsx
