+++ src/containers/NetworkPanel/NetworkDetails/ResponseView.tsx
@@ -1,7 +1,7 @@
 import React, { useMemo } from "react";
-import * as safeJson from "../../helpers/safeJson";
-import { JsonView } from "../../components/JsonView";
-import { CopyButton } from "../../components/CopyButton";
+import * as safeJson from "../../../helpers/safeJson";
+import { JsonView } from "../../../components/JsonView";
+import { CopyButton } from "../../../components/CopyButton";
 
 interface IResponseViewProps {
   response?: string;
