+++ src/containers/NetworkPanel/NetworkTable/index.test.tsx
