+++ src/containers/SearchPanel/index.tsx
@@ -1,7 +1,11 @@
 import React from "react";
 
-interface ISearchPanelProps {}
+interface ISearchPanelProps {
+  query: string;
+}
 
 export const SearchPanel = (props: ISearchPanelProps) => {
-  return <div></div>;
+  const { query } = props;
+
+  return <div>search here: {query}</div>;
 };
