+++ src/mocks/mock-chrome.ts
@@ -1,14 +1,7 @@
 import { mockRequests } from "../mocks/mock-requests";
 
-let listeners: Array<() => void> = [];
-const removeListeners = () => {
-  while (listeners.length) {
-    const listener = listeners.shift();
-    if (listener) {
-      listener();
-    }
-  }
-};
+let removeListeners: Record<string, () => void> = {};
+let searchQueryString = "";
 
 export const mockChrome = {
   devtools: {
@@ -30,17 +23,44 @@ export const mockChrome = {
             }
           };
           window.addEventListener("keydown", handleKeydown);
-          listeners.push(() => {
+          removeListeners.onRequestFinished = () =>
             window.removeEventListener("keydown", handleKeydown);
-          });
         },
         removeListener: (cb) => {
-          removeListeners();
+          removeListeners.onRequestFinished();
         },
       },
       onNavigated: {
         addListener: (cb) => {},
         removeListener: (cb) => {},
+      },
+    },
+  },
+  runtime: {
+    onMessage: {
+      addListener: (cb) => {
+        // On press key "1", add more mock requests to app
+        const handleKeydown = (e: KeyboardEvent) => {
+          let payload;
+          if (e.code === "Enter") {
+            payload = {
+              action: "nextSearchResult",
+            };
+          } else {
+            searchQueryString += e.key;
+            payload = {
+              action: "performSearch",
+              queryString: searchQueryString,
+            };
+          }
+          cb({ type: "search", payload }, "" as any, () => {});
+        };
+        window.addEventListener("keydown", handleKeydown);
+        removeListeners.onMessage = () =>
+          window.removeEventListener("keydown", handleKeydown);
+      },
+      removeListener: (cb) => {
+        removeListeners.onMessage();
       },
     },
   },
